[Fix] Do not hide FK if there is no attached navigation property (when option SkipForeignKeys is set to false)
