+++ playground/EasyDataAspNetCoreTest02/Startup.cs
@@ -64,6 +64,7 @@ namespace EasyDataBasicDemo
             app.UseEndpoints(endpoints => {
                 endpoints.MapEasyData(options => {
                     options.UseDbContext<ApplicationDbContext>(loaderOptions => {
+                        loaderOptions.SkipForeignKeys = false;
                         //loaderOptions.AddFilter(entityType => !entityType.ClrType.Name.StartsWith("Identity"));
                     });
                 });
